requirements.txt
python-telegram-bot>=21,<22
python-dotenv>=1.0
PyYAML>=6.0
pytest>=8.0
